refactor: LottosFactory와 LottoFactory  통합
