--- src/main/java/domain/factory/LottoFactory.java
@@ -1,17 +1,0 @@
-package domain.factory;
-
-import domain.Lotto;
-import global.generator.Generator;
-
-public class LottoFactory {
-    private final Generator generator;
-
-    public LottoFactory(final Generator generator) {
-        this.generator = generator;
-    }
-
-    public Lotto create() {
-        return new Lotto(generator.generate());
-    }
-
-}
